git-svn-id: http://dyncall.org/svn/dyncall/trunk@309 c71cb566-5a31-4692-9165-cf76bdc42cab

+++ bindings/R/rdyncall/src/rutils.c
@@ -22,7 +22,7 @@ SEXP r_addrval(SEXP x)
 	switch(TYPEOF(x))
 	{
 	case NILSXP: return ScalarInteger(0);
-	case EXTPTRSXP: return ScalarInteger( (int) R_ExternalPtrAddr( x ) );
+	case EXTPTRSXP: return ScalarInteger( (int) (unsigned long long) R_ExternalPtrAddr( x ) );
 	default: error("invalid type");
 	}
 	return R_NilValue;
